Refactoring code into reward_predictors module
